HBASE-30188 Upgrade hbase-server to use junit5 Part20 (#8286)
(cherry picked from commit 8ce9e3b53798b284c402875856518c0566f33949)
