+++ hbase-server/src/test/java/org/apache/hadoop/hbase/replication/TestReplicationKillMasterRS.java
@@ -17,33 +17,19 @@
  */
 package org.apache.hadoop.hbase.replication;
 
-import org.apache.hadoop.hbase.HBaseClassTestRule;
 import org.apache.hadoop.hbase.testclassification.LargeTests;
 import org.apache.hadoop.hbase.testclassification.ReplicationTests;
-import org.junit.BeforeClass;
-import org.junit.ClassRule;
-import org.junit.Test;
-import org.junit.experimental.categories.Category;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Tag;
 
-/**
- * Runs the TestReplicationKillRS test and selects the RS to kill in the master cluster Do not add
- * other tests in this class.
- */
-@Category({ ReplicationTests.class, LargeTests.class })
-public class TestReplicationKillMasterRS extends TestReplicationKillRS {
+@Tag(ReplicationTests.TAG)
+@Tag(LargeTests.TAG)
+public class TestReplicationKillMasterRS extends ReplicationKillMasterRSTestBase {
 
-  @ClassRule
-  public static final HBaseClassTestRule CLASS_RULE =
-    HBaseClassTestRule.forClass(TestReplicationKillMasterRS.class);
-
-  @BeforeClass
+  @BeforeAll
   public static void setUpBeforeClass() throws Exception {
     NUM_SLAVES1 = 2;
-    TestReplicationBase.setUpBeforeClass();
-  }
-
-  @Test
-  public void killOneMasterRS() throws Exception {
-    loadTableAndKillRS(UTIL1);
+    configureClusters(UTIL1, UTIL2);
+    startClusters();
   }
 }
